- vrtio disk is working in QEMU 1.0.1 - renamed boards with _ instead of - (needed for defining BOARD_<board-name> in code for #ifdefs) - some cleanup with boot time/debug getc/putc on i386 platform - use good cli/sti version of spinlocks in virtio - arm build cleanup

+++ trunk/oldtree/kernel/phantom/disk_struct.c
@@ -131,7 +131,7 @@ phantom_dump_superblock(phantom_disk_superblock *sb)
     printf("--------------\n");
 
     printf("sys_name          %*s\n", DISK_STRUCT_SB_SYSNAME_SIZE, sb->sys_name );
-    printf("obj spc address   0x%lx\n", sb->object_space_address );
+    printf("obj spc address   0x%lx\n", (unsigned long)sb->object_space_address );
